transform query validator to class
